* sync part of new fix to sylixos

+++ src/nng/src/sp/protocol/mqtt/mqtt_parser.c
@@ -142,11 +142,11 @@ put_var_integer(uint8_t *dest, uint32_t value)
  * Get variable integer value
  *
  * @param buf Byte array
- * @param pos how many bits rlen occupied
+ * @param pos how many bits rlen occupied (accumulated value)
  * @return Integer value
  */
 uint32_t
-get_var_integer(const uint8_t *buf, uint32_t *pos)
+get_var_integer(const uint8_t *buf, uint8_t *pos)
 {
 	uint8_t  temp;
 	uint32_t result = 0;
